# Use xxHash for the change detector